Gegner has hp bar and can be killed

+++ core/src/main/java/io/github/infotest/util/DataObjects/PlayerData.java
@@ -17,7 +17,7 @@ public class PlayerData {
     public int level;
     public double mana;
     public Rotation rotation;
-    public int gold;
+    public float gold;
 
     public static class Position {
         public double x;
